copytrace: fix copy packet_context
Since we are not filtering the fields in the packet_context anymore, we
can now use the deep copy function bt_ctf_field_copy for the
packet_context.

Refactor the code to work with that and create a function to set the
packet_context on a packet or on a stream.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>